Make long long for high numbers in limit and offset

+++ source/lib/integer_to_char.h
@@ -1,7 +1,7 @@
 #ifndef INTEGER_TO_CHAR_H
 #define INTEGER_TO_CHAR_H
 
-void integer_to_char(int value, char *str, int base);
+void integer_to_char(long long value, char *str, int base);
 
 void reverse_string(char *begin, char *end);
